v0.055 (#4)
* 1st commit to about

I still have to organize it more.

* V0.0.35

Improved the description area and added a new section.

* V0.0.38

Completed the Hero Section, Cleaned the hero code and it is now able to compress while keeping a good format. Hero section is complete.

* V0.0.4

Completed the about section with a parallax image. Need to decide the layout of the Experience section.

* Experience section complete

Completed the entire experience section. Created 1 page for each company.

* v0.055

Fixed some titles, links and deleted unnecessary code.

---------

Co-authored-by: 3DWafflez <164435244+3DWafflez@users.noreply.github.com>